test: consolidate
