+++ tests/conftest.py
@@ -1,7 +1,12 @@
-import equinox.internal as eqxi
+import jax
 import pytest
+
+
+class GetKey:
+    def __call__(self):
+        return jax.random.key(0)
 
 
 @pytest.fixture()
 def getkey():
-    return eqxi.GetKey()
+    return GetKey()
